Audit advisories, work based off of #1253 from ripienaar
