+++ server/jetstream.go
@@ -55,126 +55,6 @@ type JetStreamAccountStats struct {
 	Limits  JetStreamAccountLimits `json:"limits"`
 }
 
-// Responses to requests sent to a server from a client.
-const (
-	// OK response
-	OK = "+OK"
-	// ERR prefix response
-	ErrPrefix = "-ERR"
-
-	// JetStreamNotEnabled is returned when JetStream is not enabled.
-	JetStreamNotEnabled = "-ERR 'jetstream not enabled for account'"
-	// JetStreamBadRequest is returned when the request could not be properly parsed.
-	JetStreamBadRequest = "-ERR 'bad request'"
-)
-
-// Request API subjects for JetStream.
-const (
-	// JetStreamEnabled allows a user to dynamically check if JetStream is enabled for an account.
-	// Will return +OK on success, otherwise will timeout.
-	JetStreamEnabled = "$JS.ENABLED"
-
-	// JetStreamInfo is for obtaining general information about JetStream for this account.
-	// Will return JSON response.
-	JetStreamInfo = "$JS.INFO"
-
-	// JetStreamCreateTemplate is the endpoint to create new stream templates.
-	// Will return +OK on success and -ERR on failure.
-	JetStreamCreateTemplate  = "$JS.TEMPLATE.*.CREATE"
-	JetStreamCreateTemplateT = "$JS.TEMPLATE.%s.CREATE"
-
-	// JetStreamListTemplates is the endpoint to list all stream templates for this account.
-	// Will return json list of string on success and -ERR on failure.
-	JetStreamListTemplates = "$JS.TEMPLATES.LIST"
-
-	// JetStreamTemplateInfo is for obtaining general information about a named stream template.
-	// Will return JSON response.
-	JetStreamTemplateInfo  = "$JS.TEMPLATE.*.INFO"
-	JetStreamTemplateInfoT = "$JS.TEMPLATE.%s.INFO"
-
-	// JetStreamDeleteTemplate is the endpoint to delete stream templates.
-	// Will return +OK on success and -ERR on failure.
-	JetStreamDeleteTemplate  = "$JS.TEMPLATE.*.DELETE"
-	JetStreamDeleteTemplateT = "$JS.TEMPLATE.%s.DELETE"
-
-	// JetStreamCreateStream is the endpoint to create new streams.
-	// Will return +OK on success and -ERR on failure.
-	JetStreamCreateStream  = "$JS.STREAM.*.CREATE"
-	JetStreamCreateStreamT = "$JS.STREAM.%s.CREATE"
-
-	// JetStreamListStreams is the endpoint to list all streams for this account.
-	// Will return json list of string on success and -ERR on failure.
-	JetStreamListStreams = "$JS.STREAM.LIST"
-
-	// JetStreamStreamInfo is for obtaining general information about a named stream.
-	// Will return JSON response.
-	JetStreamStreamInfo  = "$JS.STREAM.*.INFO"
-	JetStreamStreamInfoT = "$JS.STREAM.%s.INFO"
-
-	// JetStreamDeleteStream is the endpoint to delete streams.
-	// Will return +OK on success and -ERR on failure.
-	JetStreamDeleteStream  = "$JS.STREAM.*.DELETE"
-	JetStreamDeleteStreamT = "$JS.STREAM.%s.DELETE"
-
-	// JetStreamPurgeStream is the endpoint to purge streams.
-	// Will return +OK on success and -ERR on failure.
-	JetStreamPurgeStream  = "$JS.STREAM.*.PURGE"
-	JetStreamPurgeStreamT = "$JS.STREAM.%s.PURGE"
-
-	// JetStreamDeleteMsg is the endpoint to delete messages from a stream.
-	// Will return +OK on success and -ERR on failure.
-	JetStreamDeleteMsg  = "$JS.STREAM.*.MSG.DELETE"
-	JetStreamDeleteMsgT = "$JS.STREAM.%s.MSG.DELETE"
-
-	// JetStreamCreateConsumer is the endpoint to create durable consumers for streams.
-	// You need to include the stream and consumer name in the subject.
-	// Will return +OK on success and -ERR on failure.
-	JetStreamCreateConsumer  = "$JS.STREAM.*.CONSUMER.*.CREATE"
-	JetStreamCreateConsumerT = "$JS.STREAM.%s.CONSUMER.%s.CREATE"
-
-	// JetStreamCreateEphemeralConsumer is the endpoint to create ephemeral consumers for streams.
-	// Will return +OK <consumer name> on success and -ERR on failure.
-	JetStreamCreateEphemeralConsumer  = "$JS.STREAM.*.EPHEMERAL.CONSUMER.CREATE"
-	JetStreamCreateEphemeralConsumerT = "$JS.STREAM.%s.EPHEMERAL.CONSUMER.CREATE"
-
-	// JetStreamConsumers is the endpoint to list all consumers for the stream.
-	// Will return json list of string on success and -ERR on failure.
-	JetStreamConsumers  = "$JS.STREAM.*.CONSUMERS"
-	JetStreamConsumersT = "$JS.STREAM.%s.CONSUMERS"
-
-	// JetStreamConsumerInfo is for obtaining general information about a consumer.
-	// Will return JSON response.
-	JetStreamConsumerInfo  = "$JS.STREAM.*.CONSUMER.*.INFO"
-	JetStreamConsumerInfoT = "$JS.STREAM.%s.CONSUMER.%s.INFO"
-
-	// JetStreamDeleteConsumer is the endpoint to delete consumers.
-	// Will return +OK on success and -ERR on failure.
-	JetStreamDeleteConsumer  = "$JS.STREAM.*.CONSUMER.*.DELETE"
-	JetStreamDeleteConsumerT = "$JS.STREAM.%s.CONSUMER.%s.DELETE"
-
-	// JetStreamAckT is the template for the ack message stream coming back from an consumer
-	// when they ACK/NAK, etc a message.
-	JetStreamAckT = "$JS.ACK.%s.%s"
-
-	// JetStreamRequestNextT is the prefix for the request next message(s) for a consumer in worker/pull mode.
-	JetStreamRequestNextT = "$JS.STREAM.%s.CONSUMER.%s.NEXT"
-
-	// JetStreamMsgBySeqT is the template for direct requests for a message by its stream sequence number.
-	JetStreamMsgBySeqT = "$JS.STREAM.%s.MSG.BYSEQ"
-
-	// JetStreamAdvisoryPrefix is a prefix for all JetStream advisories
-	JetStreamAdvisoryPrefix = "$JS.EVENT.ADVISORY"
-
-	// JetStreamMetricPrefix is a prefix for all JetStream metrics
-	JetStreamMetricPrefix = "$JS.EVENT.METRIC"
-
-	// JetStreamMetricConsumerAckPre is a metric containing ack latency
-	JetStreamMetricConsumerAckPre = JetStreamMetricPrefix + ".CONSUMER_ACK"
-
-	// JetStreamAdvisoryConsumerMaxDeliveryExceedPre is a notification published when a message exceeds its delivery threshold
-	JetStreamAdvisoryConsumerMaxDeliveryExceedPre = JetStreamAdvisoryPrefix + ".MAX_DELIVERIES"
-)
-
 // This is for internal accounting for JetStream for this server.
 type jetStream struct {
 	mu            sync.RWMutex
@@ -203,27 +83,6 @@ type jsAccount struct {
 	streams       map[string]*Stream
 	templates     map[string]*StreamTemplate
 	store         TemplateStore
-}
-
-// For easier handling of exports and imports.
-var allJsExports = []string{
-	JetStreamEnabled,
-	JetStreamInfo,
-	JetStreamCreateTemplate,
-	JetStreamListTemplates,
-	JetStreamTemplateInfo,
-	JetStreamDeleteTemplate,
-	JetStreamCreateStream,
-	JetStreamListStreams,
-	JetStreamStreamInfo,
-	JetStreamDeleteStream,
-	JetStreamPurgeStream,
-	JetStreamDeleteMsg,
-	JetStreamCreateConsumer,
-	JetStreamCreateEphemeralConsumer,
-	JetStreamConsumers,
-	JetStreamConsumerInfo,
-	JetStreamDeleteConsumer,
 }
 
 // EnableJetStream will enable JetStream support on this server with the given configuration.
@@ -281,55 +140,7 @@ func (s *Server) EnableJetStream(config *JetStreamConfig) error {
 	}
 
 	// Setup our internal subscriptions.
-	if _, err := s.sysSubscribe(JetStreamEnabled, s.isJsEnabledRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamInfo, s.jsAccountInfoRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamCreateTemplate, s.jsCreateTemplateRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamListTemplates, s.jsTemplateListRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamTemplateInfo, s.jsTemplateInfoRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamDeleteTemplate, s.jsTemplateDeleteRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamCreateStream, s.jsCreateStreamRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamListStreams, s.jsStreamListRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamStreamInfo, s.jsStreamInfoRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamDeleteStream, s.jsStreamDeleteRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamPurgeStream, s.jsStreamPurgeRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamDeleteMsg, s.jsMsgDeleteRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamCreateConsumer, s.jsCreateConsumerRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamCreateEphemeralConsumer, s.jsCreateEphemeralConsumerRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamConsumers, s.jsConsumersRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamConsumerInfo, s.jsConsumerInfoRequest); err != nil {
-		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
-	}
-	if _, err := s.sysSubscribe(JetStreamDeleteConsumer, s.jsConsumerDeleteRequest); err != nil {
+	if err := s.setJetStreamExportSubs(); err != nil {
 		return fmt.Errorf("Error setting up internal jetstream subscriptions: %v", err)
 	}
 
@@ -986,474 +797,6 @@ func (js *jetStream) releaseResources(limits *JetStreamAccountLimits) error {
 		js.storeReserved -= limits.MaxStore
 	}
 	return nil
-}
-
-// Request to check if jetstream is enabled.
-func (s *Server) isJsEnabledRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, OK)
-	} else {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-	}
-}
-
-// Request for current usage and limits for this account.
-func (s *Server) jsAccountInfoRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	stats := c.acc.JetStreamUsage()
-	b, err := json.MarshalIndent(stats, "", "  ")
-	if err != nil {
-		return
-	}
-	s.sendInternalAccountMsg(c.acc, reply, b)
-}
-
-// Request to create a new template.
-func (s *Server) jsCreateTemplateRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	var cfg StreamTemplateConfig
-	if err := json.Unmarshal(msg, &cfg); err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	templateName := subjectToken(subject, 2)
-	if templateName != cfg.Name {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr("template name in subject does not match request"))
-		return
-	}
-
-	var response = OK
-	if _, err := c.acc.AddStreamTemplate(&cfg); err != nil {
-		response = protoErr(err)
-	}
-	s.sendInternalAccountMsg(c.acc, reply, response)
-}
-
-// Request for the list of all templates.
-func (s *Server) jsTemplateListRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	var names []string
-	ts := c.acc.Templates()
-	for _, t := range ts {
-		t.mu.Lock()
-		name := t.Name
-		t.mu.Unlock()
-		names = append(names, name)
-	}
-	b, err := json.MarshalIndent(names, "", "  ")
-	if err != nil {
-		return
-	}
-	s.sendInternalAccountMsg(c.acc, reply, b)
-}
-
-// Request for information about a stream template.
-func (s *Server) jsTemplateInfoRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	if len(msg) != 0 {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	name := subjectToken(subject, 2)
-	t, err := c.acc.LookupStreamTemplate(name)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-	t.mu.Lock()
-	cfg := t.StreamTemplateConfig.deepCopy()
-	streams := t.streams
-	t.mu.Unlock()
-	si := &StreamTemplateInfo{
-		Config:  cfg,
-		Streams: streams,
-	}
-	b, err := json.MarshalIndent(si, "", "  ")
-	if err != nil {
-		return
-	}
-	s.sendInternalAccountMsg(c.acc, reply, b)
-}
-
-// Request to delete a stream template.
-func (s *Server) jsTemplateDeleteRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	if len(msg) != 0 {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	name := subjectToken(subject, 2)
-	err := c.acc.DeleteStreamTemplate(name)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-	s.sendInternalAccountMsg(c.acc, reply, OK)
-}
-
-// Request to create a stream.
-func (s *Server) jsCreateStreamRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	var cfg StreamConfig
-	if err := json.Unmarshal(msg, &cfg); err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	streamName := subjectToken(subject, 2)
-	if streamName != cfg.Name {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr("stream name in subject does not match request"))
-		return
-	}
-
-	var response = OK
-	if _, err := c.acc.AddStream(&cfg); err != nil {
-		response = protoErr(err)
-	}
-	s.sendInternalAccountMsg(c.acc, reply, response)
-}
-
-// Request for the list of all streams.
-func (s *Server) jsStreamListRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	var names []string
-	msets := c.acc.Streams()
-	for _, mset := range msets {
-		names = append(names, mset.Name())
-	}
-	b, err := json.MarshalIndent(names, "", "  ")
-	if err != nil {
-		return
-	}
-	s.sendInternalAccountMsg(c.acc, reply, b)
-}
-
-// Request for information about a stream.
-func (s *Server) jsStreamInfoRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	if len(msg) != 0 {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	name := subjectToken(subject, 2)
-	mset, err := c.acc.LookupStream(name)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-	msi := StreamInfo{
-		State:  mset.State(),
-		Config: mset.Config(),
-	}
-	b, err := json.MarshalIndent(msi, "", "  ")
-	if err != nil {
-		return
-	}
-	s.sendInternalAccountMsg(c.acc, reply, b)
-}
-
-// Request to delete a stream.
-func (s *Server) jsStreamDeleteRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	if len(msg) != 0 {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	name := subjectToken(subject, 2)
-	mset, err := c.acc.LookupStream(name)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-	var response = OK
-	if err := mset.Delete(); err != nil {
-		response = protoErr(err)
-	}
-	s.sendInternalAccountMsg(c.acc, reply, response)
-}
-
-// Request to delete a message.
-// This expects a stream sequence number as the msg body.
-func (s *Server) jsMsgDeleteRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	if len(msg) == 0 {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	name := subjectToken(subject, 2)
-	mset, err := c.acc.LookupStream(name)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-	var response = OK
-	seq, _ := strconv.Atoi(string(msg))
-	if !mset.EraseMsg(uint64(seq)) {
-		response = protoErr(fmt.Sprintf("sequence [%d] not found", seq))
-	}
-	s.sendInternalAccountMsg(c.acc, reply, response)
-}
-
-// Request to purge a stream.
-func (s *Server) jsStreamPurgeRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	if len(msg) != 0 {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	name := subjectToken(subject, 2)
-	mset, err := c.acc.LookupStream(name)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-
-	mset.Purge()
-	s.sendInternalAccountMsg(c.acc, reply, OK)
-}
-
-// Request to create a durable consumer.
-func (s *Server) jsCreateConsumerRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	var req CreateConsumerRequest
-	if err := json.Unmarshal(msg, &req); err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	streamName := subjectToken(subject, 2)
-	if streamName != req.Stream {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr("stream name in subject does not match request"))
-		return
-	}
-	stream, err := c.acc.LookupStream(req.Stream)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-	// Now check we do not have a durable.
-	if req.Config.Durable == _EMPTY_ {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr("consumer expected to be durable but a durable name was not set"))
-		return
-	}
-	consumerName := subjectToken(subject, 4)
-	if consumerName != req.Config.Durable {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr("consumer name in subject does not match durable name in request"))
-		return
-	}
-	var response = OK
-	if _, err := stream.AddConsumer(&req.Config); err != nil {
-		response = protoErr(err)
-	}
-	s.sendInternalAccountMsg(c.acc, reply, response)
-}
-
-// Request to create an ephemeral consumer.
-func (s *Server) jsCreateEphemeralConsumerRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	var req CreateConsumerRequest
-	if err := json.Unmarshal(msg, &req); err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	streamName := subjectToken(subject, 2)
-	if streamName != req.Stream {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr("stream name in subject does not match request"))
-		return
-	}
-	stream, err := c.acc.LookupStream(req.Stream)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-	// Now check we do not have a durable.
-	if req.Config.Durable != _EMPTY_ {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr("consumer expected to be ephemeral but a durable name was set"))
-		return
-	}
-	var response = OK
-	if o, err := stream.AddConsumer(&req.Config); err != nil {
-		response = protoErr(err)
-	} else if !o.isDurable() {
-		// If the consumer is ephemeral add in the name
-		response = OK + " " + o.Name()
-	}
-	s.sendInternalAccountMsg(c.acc, reply, response)
-}
-
-// Request for the list of all consumers.
-func (s *Server) jsConsumersRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	if len(msg) != 0 {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	name := subjectToken(subject, 2)
-	mset, err := c.acc.LookupStream(name)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-	var onames []string
-	obs := mset.Consumers()
-	for _, o := range obs {
-		onames = append(onames, o.Name())
-	}
-	b, err := json.MarshalIndent(onames, "", "  ")
-	if err != nil {
-		return
-	}
-	s.sendInternalAccountMsg(c.acc, reply, b)
-}
-
-// Request for information about an consumer.
-func (s *Server) jsConsumerInfoRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	if len(msg) != 0 {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	stream := subjectToken(subject, 2)
-	mset, err := c.acc.LookupStream(stream)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-	consumer := subjectToken(subject, 4)
-	obs := mset.LookupConsumer(consumer)
-	if obs == nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr("consumer not found"))
-		return
-	}
-	info := obs.Info()
-	b, err := json.MarshalIndent(info, "", "  ")
-	if err != nil {
-		return
-	}
-	s.sendInternalAccountMsg(c.acc, reply, b)
-}
-
-// Request to delete an Consumer.
-func (s *Server) jsConsumerDeleteRequest(sub *subscription, c *client, subject, reply string, msg []byte) {
-	if c == nil || c.acc == nil {
-		return
-	}
-	if !c.acc.JetStreamEnabled() {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamNotEnabled)
-		return
-	}
-	if len(msg) != 0 {
-		s.sendInternalAccountMsg(c.acc, reply, JetStreamBadRequest)
-		return
-	}
-	stream := subjectToken(subject, 2)
-	mset, err := c.acc.LookupStream(stream)
-	if err != nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr(err))
-		return
-	}
-	consumer := subjectToken(subject, 4)
-	obs := mset.LookupConsumer(consumer)
-	if obs == nil {
-		s.sendInternalAccountMsg(c.acc, reply, protoErr("consumer not found"))
-		return
-	}
-	var response = OK
-	if err := obs.Delete(); err != nil {
-		response = protoErr(err)
-	}
-	s.sendInternalAccountMsg(c.acc, reply, response)
 }
 
 const (
